Support controllers with unimplemented properties
Some virtual controllers, notably ones created with Steam Input, do not
implement the function that returns a JoystickId, amongst other
properties. As such, attempting to configure such a controller would
result in Exceptions being raised, and general inability of the Mod to
even initialize, let alone configure controls assigned to axes governed
by such virtual controllers.

We can work around this by replacing references to the JoystickId with
references to the index in the array of joysticks provided to us by
SharpDX, by removing all references to the JoystickId from all
logging and lookup functions, and by doing a better job of guarding
against exceptions when logging properties for controllers on startup.