Support virtual axes on the same physical axis
Players may have a physical "lever" or "throttle" style joystick,
and prefer to split the same physical lever into multiple ranges in
order to cover multiple train controls. For instance, this can be used
to assign the lower range to the independent brake and the upper range
to the train brake. This way, a single physical movement can move
through multiple in-game brake levers without additional hardware.

With this in mind, this patch adds "min" and "max" virtual range
selectors into the config to allow these axes to be split by the player.
The player can read through the log as they move the lever to understand
the raw values of the input axis and input these into the configuration.